Init and Command base class
