+++ ContraControl/dev/Controller/commands.h
@@ -1,0 +1,8 @@
+#ifndef COMMANDS_INCLUSION_H
+#define COMMANDS_INCLUSION_H
+
+
+#include "Commands/command.h"
+#include "Commands/initalize.h"
+
+#endif // !COMMANDS_INCLUSION_H
