+++ ContraControl/dev/Model/devices.h
@@ -1,0 +1,2 @@
+#include "devices/device.h"
+#include "devices/switch_device.h"
